fix(gateway): fix bigint type mismatch and api_calls schema in generate/proxy routes
Credit ledger SUM(delta) returns bigint, causing type conflict with integer
parameter. Added ::int cast. Also fixed api_calls INSERT in generate route
to match actual table schema (session_id, provider, credits_charged columns).

+++ gateway/src/routes/proxy.ts
@@ -25,7 +25,7 @@ proxy.post('/v1/chat/completions', embedTokenAuth, async (c) => {
       }
       await db.query(
         `INSERT INTO subscriptions.credit_ledger (user_id, delta, balance_after, reason, app_id)
-         VALUES ($1, $2, (SELECT COALESCE(SUM(delta), 0) + $2 FROM subscriptions.credit_ledger WHERE user_id = $1), 'api_call', $3)`,
+         VALUES ($1, $2, (SELECT COALESCE(SUM(delta), 0)::int + $2 FROM subscriptions.credit_ledger WHERE user_id = $1), 'api_call', $3)`,
         [userId, -creditsPerCall, appId],
       )
       creditsCharged = creditsPerCall
